refactor(/payment, /toss): 결제 client 관련 파일들의 위치 수정
