+++ src/main/java/roomescape/payment/client/toss/TossClientErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.common;
+package roomescape.payment.client.toss;
 
 public record TossClientErrorResponse(String code,
                                       String message) {
